+++ src/Commands/music/play/spotify.ts
@@ -4,7 +4,9 @@ import {Queue} from "../../../Struct/Queue";
 import {TrackType} from "../../../Interface/Track";
 import {Album, Artist, parse, Playlist, Track as SpotifyUriTrack} from "spotify-uri";
 import {Track as SpotifyApiTrack} from "spotify-api.js";
+import {isUrl} from "../../../utils/function";
 
+//TODO change everything because that is not optimal
 module.exports = async (interaction: ChatInputCommandInteraction, client: Bot, queue: Queue) => {
 
     const query = interaction.options.getString('name', true);
@@ -15,12 +17,11 @@ module.exports = async (interaction: ChatInputCommandInteraction, client: Bot, q
 
     await interaction.reply({content: `Searching music ...`, ephemeral: true});
 
-    try {
+    if(isUrl(query)) {
         const url = new URL(query);
 
         if (url.hostname.includes('spotify') || url.protocol.includes('spotify')) {
             const data = parse(query);
-
             if (data instanceof SpotifyUriTrack) {
                 const track = await client.player.spotifyClient.searchTrack(data.id);
                 if (!track) return interaction.followUp({content: 'No results found!', ephemeral: true});
@@ -28,6 +29,7 @@ module.exports = async (interaction: ChatInputCommandInteraction, client: Bot, q
                 const ytTrack = await client.player.searchYoutubeTrack(`${track.name} ${track.artists[0].name}`);
                 if (!ytTrack) return interaction.followUp({content: 'No results found!', ephemeral: true});
                 ytTrack.type = TrackType.SPOTIFY;
+                ytTrack.resource.volume?.setVolume(0.25);
                 queue.addTrack(ytTrack);
                 await interaction.followUp({content: `Add ${ytTrack.title} to Queue`, ephemeral: false});
 
@@ -36,10 +38,14 @@ module.exports = async (interaction: ChatInputCommandInteraction, client: Bot, q
                 if (!album?.tracks) return interaction.followUp({content: 'No results found!', ephemeral: true});
                 const listTrack : any[] = [];
                 for (const track of album.tracks) {
-                    const ytTrack = await client.player.searchYoutubeTrack(`${track.name} ${track.artists[0].name}`);
-                    if (ytTrack) {
-                        ytTrack.type = TrackType.SPOTIFY;
-                        listTrack.push(ytTrack);
+                    if (track?.name) {
+                        console.log(track.name);
+                        const ytTrack = await client.player.searchYoutubeTrack(`${track.name} ${track.artists[0].name}`);
+                        if (ytTrack) {
+                            ytTrack.type = TrackType.SPOTIFY;
+                            ytTrack.resource.volume?.setVolume(0.25);
+                            listTrack.push(ytTrack);
+                        }
                     }
                 }
                 queue.addTracks(listTrack);
@@ -60,27 +66,32 @@ module.exports = async (interaction: ChatInputCommandInteraction, client: Bot, q
                 })
                 search.type = TrackType.SPOTIFY;
                 queue.addTrack(search);
+                search.resource.volume?.setVolume(0.25);
                 await interaction.followUp({content: `Add ${search.title} to Queue`, ephemeral: false});
 
             } else if (data instanceof Playlist) {
                 const playlist = await client.player.spotifyClient.searchPlaylist(data.id);
                 if (!playlist?.tracks) return interaction.followUp({content: 'No results found!', ephemeral: true});
 
+                const listTrack : any[] = [];
                 for (const track of playlist.tracks) {
                     if (track.track instanceof SpotifyApiTrack) {
                         const ytTrack = await client.player.searchYoutubeTrack(`${track.track?.name} ${track.track?.artists[0].name}`);
-                        if (!ytTrack) return interaction.followUp({content: 'No results found!', ephemeral: true});
-                        ytTrack.type = TrackType.SPOTIFY;
-                        queue.addTrack(ytTrack);
-                        await interaction.followUp({content: `Add ${ytTrack.title} to Queue`, ephemeral: false});
+                        if (ytTrack) {
+                            ytTrack.type = TrackType.SPOTIFY;
+                            ytTrack.resource.volume?.setVolume(0.25);
+                            listTrack.push(ytTrack);
+                        }
                     } else if (track.track) {
                         const ytTrack = await client.player.searchYoutubeTrack(`${track.track?.name} ${track.track?.show?.name}`);
-                        if (!ytTrack) return interaction.followUp({content: 'No results found!', ephemeral: true});
-                        ytTrack.type = TrackType.SPOTIFY;
-                        queue.addTrack(ytTrack);
-                        await interaction.followUp({content: `Add ${ytTrack.title} to Queue`, ephemeral: false});
+                        if (ytTrack) {
+                            ytTrack.type = TrackType.SPOTIFY;
+                            ytTrack.resource.volume?.setVolume(0.25);
+                            listTrack.push(ytTrack);
+                        }
                     }
                 }
+                await interaction.followUp({content: `Add ${playlist.name} to Queue - ${listTrack.length} tracks`, ephemeral: false});
 
             } else {
                 return interaction.followUp({content: 'No results found!', ephemeral: true});
@@ -89,17 +100,18 @@ module.exports = async (interaction: ChatInputCommandInteraction, client: Bot, q
         } else {
             return interaction.followUp({content: 'Invalid link please use official spotify or open issue on the github to inform my creator', ephemeral: true});
         }
-    } catch (e) {
-
+    } else {
         const data = await client.player.spotifyClient.search(query);
 
         if (!data) return interaction.followUp({content: 'No results found!', ephemeral: true});
 
-         if (data.tracks) {
+        if (data.tracks) {
             const track = data.tracks[0];
+            if (!track?.name) return interaction.followUp({content: 'No results found!', ephemeral: true});
             const ytTrack = await client.player.searchYoutubeTrack(`${track.name} ${track.artists[0].name}`);
             if (!ytTrack) return interaction.followUp({content: 'No results found!', ephemeral: true});
             ytTrack.type = TrackType.SPOTIFY;
+            ytTrack.resource.volume?.setVolume(0.25);
             queue.addTrack(ytTrack);
 
             await interaction.followUp({content: `Added ${ytTrack.title} to the queue!`, ephemeral: false});
@@ -115,7 +127,6 @@ module.exports = async (interaction: ChatInputCommandInteraction, client: Bot, q
     if (!queue.playing) {
         setTimeout(() => {
             queue.play();
-            queue.setVolume(50);
         }, 1000)
     }
 }