+++ src/Commands/music/queue.ts
@@ -5,15 +5,22 @@ import {Bot} from "../../Struct/Bot";
 
 module.exports = async (interaction: ChatInputCommandInteraction, client: Bot, queue: Queue) => {
 
-    const track : Track[] = queue.getQueue();
+    const page = interaction.options.getInteger('page', false) || 1;
 
-    if (track.length === 0) return interaction.reply({content: `No track in queue`, ephemeral: true});
+    const queueList : Track[] = queue.getQueue();
+
+    if (queueList.length === 0) return interaction.reply({content: `No track in queue`, ephemeral: true});
+
+    if ((page-1)*10 > queueList.length) return interaction.reply({content: 'Page not found', ephemeral: true});
 
     let message = '';
 
-    track.forEach((track, index) => {
-        message += `${index + 1}. **${track.type}** -  ${track.title} - **${track.channelTitle}**\n`
-    })
+    for (let i = (page-1)*10; i < (page*10 > queueList.length ? queueList.length : page*10); i++) {
+        if (queueList[i]) {
+            message += `${i + 1}. **${queueList[i].type}** -  ${queueList[i].title} - **${queueList[i].channelTitle}**\n`
 
-    await interaction.reply({content: message, ephemeral: true});
+        }
+    }
+
+    await interaction.reply({content: message.length < 1 ? 'queue is Empty' : message, ephemeral: message.length < 1});
 }