+++ src/Events/core/onReady.ts
@@ -1,5 +1,6 @@
 import {Events} from "discord.js";
 import {Bot} from "../../Struct/Bot";
+import {deleteFile} from "../../utils/function";
 
 module.exports = {
     name: Events.ClientReady,
@@ -14,5 +15,10 @@ module.exports = {
                 client.player.createQueue(guild.id);
             })
         })
+
+        await deleteFile('youtube.json');
+        setInterval(() => {
+            deleteFile('youtube.json');
+        }, 1000 * 60 * 60 * 24 * 2);
     }
 }