+++ changelog.md
@@ -1,7 +1,7 @@
 ## 0.0.2
  - initial commit
  - basic bot functionality focused on music (and openAi)
- - added twitch, youtube and spotify support
+ - added twitch, YouTube and spotify support
 
 ## 0.1.0
  - add autocomplete support for every search command (youtube, spotify, twitch)